[Rename#23]: DTO_File_Change_Structure
