+++ src/main/java/org/crews/controller/MemberController.java
@@ -4,8 +4,8 @@ import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
-import org.crews.dto.MemberRequest;
-import org.crews.dto.MemberResponse;
+import org.crews.dto.request.MemberRequest;
+import org.crews.dto.response.MemberResponse;
 import org.crews.dto.response.InterestingResponseDto;
 import org.crews.dto.response.MyProfileResponse;
 import org.crews.dto.response.MyinfoResponse;
@@ -75,6 +75,12 @@ public class MemberController {
 
         return new ResponseEntity<>(HttpStatus.OK);
     }
+    @GetMapping("/signup/validate-email")
+    public ResponseEntity<String> validateEmail() {
+
+        return ResponseEntity.ok("ok");
+    }
+
     @GetMapping("/me")
     public ResponseEntity<MyinfoResponse> getMyinfo() {
         // TODO: JWT ( role , email , expired )
