+++ src/main/java/org/crews/dto/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package org.crews.dto;
+package org.crews.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
