+++ src/main/java/org/crews/jwt/JWTFilter.java
@@ -6,7 +6,7 @@ import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.extern.slf4j.Slf4j;
-import org.crews.dto.MemberDetails;
+import org.crews.dto.request.MemberDetails;
 import org.crews.model.Member;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
