+++ src/main/java/org/crews/jwt/LoginFilter.java
@@ -7,8 +7,8 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.crews.dto.LoginRequest;
-import org.crews.dto.MemberDetails;
+import org.crews.dto.request.LoginRequest;
+import org.crews.dto.request.MemberDetails;
 import org.crews.model.RefreshEntity;
 import org.crews.repository.RefreshRepository;
 import org.springframework.http.HttpStatus;
