+++ src/main/java/org/crews/service/AccountService.java
@@ -2,7 +2,7 @@ package org.crews.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.crews.dto.AccountLinkRequest;
+import org.crews.dto.request.AccountLinkRequest;
 import org.crews.dto.MemberIdDto;
 import org.crews.dto.core.AccountIssuedResponse;
 import org.crews.dto.core.AccountOneResponse;
