+++ src/main/java/org/crews/service/MemberServiceImpl.java
@@ -4,8 +4,8 @@ import io.jsonwebtoken.ExpiredJwtException;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.crews.dto.MemberRequest;
-import org.crews.dto.MemberResponse;
+import org.crews.dto.request.MemberRequest;
+import org.crews.dto.response.MemberResponse;
 import org.crews.dto.core.AccountResponseDto;
 import org.crews.dto.core.CIRequest;
 import org.crews.dto.core.MemberToCoreDto;
